backend/requirements.txt
langchain
pymongo
langchain-openai
langchain-community
pypdf
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.10.0
motor==3.3.2
pydantic>=1.10.0,<2.0.0
python-multipart==0.0.6
aiohttp==3.9.3
pytest==8.0.0
python-jose[cryptography]==3.3.0
flask==3.0.2
flask-cors==4.0.0
